requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.3
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
easydict==1.13
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gdown==5.2.0
grpcio==1.75.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.6
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.19.0
onnxruntime-gpu==1.23.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
prettytable==3.16.0
protobuf==6.32.1
psutil==7.1.0
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
soupsieve==2.8
stringzilla==4.1.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.10.4
torch==2.8.0
torchreid==0.2.5
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.206
ultralytics-thop==2.0.17
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
